# internal/interface/http/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: Stock Service API
  version: "1.0.0"
  description: API for managing inventory, stores, promotions, warehouses, and related resources.
paths:
...
  /api/v1/stores:
    get:
      summary: List stores with optional filters
      tags: [Store]
      parameters:
        - name: user_id
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"
    post:
      summary: Create a new store
      tags: [Store]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, store_name]
              properties:
                user_id:    { type: integer, format: int64 }
                store_name: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/stores/{id}:
    get:
      summary: Get store by ID
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update store name
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a store
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "deleted" }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stores/{id}/bulk-sale:
    post:
      summary: Toggle bulk sale for a store
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
  /api/v1/stores/{id}/commission:
    post:
      summary: Toggle commission for a store
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
  /api/v1/stores/{id}/contact:
    put:
      summary: Update store contact information
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact_phone: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
  /api/v1/stores/{id}/profile:
    put:
      summary: Update store profile (address, media assets)
      tags: [Store]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address_id:   { type: integer, format: int64, nullable: true }
                media_assets: { type: object, additionalProperties: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/warehouse-links:
    post:
      summary: Create a link between a store and a warehouse
      tags: [Warehouse Link]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [store_id, warehouse_id]
              properties:
                store_id:     { type: integer, format: int64 }
                warehouse_id: { type: integer, format: int64 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseLinkResponse"
  /api/v1/warehouse-links/{id}/relation:
    put:
      summary: Change the relation type of a warehouse link
      tags: [Warehouse Link]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [relation_type]
              properties:
                relation_type: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseLinkResponse"
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    StoreResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        store_name:
          type: string
        status:
          type: string
        address_id:
          type: integer
          format: int64
          nullable: true
        contact_phone:
          type: string
          nullable: true
        media_assets:
          type: object
          additionalProperties: true
        is_commission_applicable:
          type: boolean
        is_bulk_sale_enabled:
          type: boolean
        created_at:
          type: string
    StoreListResponse:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/StoreResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    WarehouseLinkResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        store_id:
          type: integer
          format: int64
        warehouse_id:
          type: integer
          format: int64
        relation_type:
          type: string
...